Document the InterpreterOptions class and the InterpreterOptions parameter to InterpreterBuilder as being experimental.
Rationale: All the methods of InterpreterOptions are already experimental,
so there's currently *nothing* that you can usefully do with this
class without using experimental features. Hence I think it is best to
also make the class itself experimental, and likewise for its use in
InterpreterBuilder, to preserves implementation flexibility for TF Lite.
PiperOrigin-RevId: 431544142